pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinker"
version = "0.1.0"
description = "AI-powered observability and incident response agent"
requires-python = ">=3.12"
dependencies = [
    # Core
    "anthropic>=0.25",
    "mcp>=1.0",
    "pydantic-settings>=2",
    "structlog>=24",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    # Server
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pyjwt[crypto]>=2.8",
    # CLI + UI
    "typer>=0.12",
    "rich>=13",
    # Slack
    "slack-bolt>=1.18",
    # AWS
    "boto3>=1.34",
    # GCP
    "google-cloud-monitoring>=2.19",
    "google-cloud-logging>=3.9",
    # Azure
    "azure-identity>=1.17",
    "azure-monitor-query>=1.3",
    # Elasticsearch / OpenSearch
    "elasticsearch>=8",
    # Code tools
    "gitpython>=3.1",
    "pygithub>=2.3",
    # Scheduling
    "apscheduler>=3.10",
]

[project.scripts]
tinker                = "tinker.interfaces.cli:app"
tinker-server         = "tinker.server.app:main"
tinker-cloudwatch-mcp = "tinker.mcp_servers.cloudwatch_server:main"
tinker-elastic-mcp    = "tinker.mcp_servers.elastic_server:main"
tinker-github-mcp     = "tinker.mcp_servers.github_server:main"
tinker-azure-mcp      = "tinker.mcp_servers.azure_server:main"
tinker-grafana-mcp    = "tinker.mcp_servers.grafana_server:main"
tinker-datadog-mcp    = "tinker.mcp_servers.datadog_server:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "moto[cloudwatch,logs]>=5",
    "vcrpy>=6",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tinker"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
